Misc update.

+++ src/map/skill.c
@@ -16824,7 +16824,7 @@ int skill_vfcastfix(struct block_list *bl, double time, uint16 skill_id, uint16 
 		time = time * (1 - sqrt(((float)(status_get_dex(bl) * 2 + status_get_int(bl)) / battle_config.vcast_stat_scale)));
 
 	time = time * (1 - (float)min(reduce_ct_r, 100) / 100);
-	time = max(time + (1 - (float)min(fixcast_r, 100) / 100) * max(fixed, 0), 0); //Underflow checking/capping
+	time = max(time + (1 - (double)min(fixcast_r, 100) / 100) * max(fixed, 0), 0); //Underflow checking/capping
 
 	return (int)time;
 }
